+++ src/CodeMap/LeetspeakMap.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 
 namespace Stadly\PasswordPolice\CodeMap;
 
-use Stadly\PasswordPolice\CharTree;
 use Stadly\PasswordPolice\CodeMap;
-use Stadly\PasswordPolice\Formatter\LengthFilter;
-use Stadly\PasswordPolice\Formatter\Truncator;
 
 final class LeetspeakMap implements CodeMap
 {
@@ -41,11 +38,6 @@ final class LeetspeakMap implements CodeMap
     private $codeMap = [];
 
     /**
-     * @var Truncator[] Formatter for extracting the first characters.
-     */
-    private $charExtractors;
-
-    /**
      * @param bool $encode Whether the map should encode or decode Leetspeak.
      */
     public function __construct($encode = false)
@@ -61,35 +53,6 @@ final class LeetspeakMap implements CodeMap
                 }
             }
         }
-        foreach (array_keys($this->codeMap + [1 => true]) as $length) {
-            $truncator = new Truncator($length);
-            $truncator->setNext(new LengthFilter($length, $length));
-            $this->charExtractors[$length] = $truncator;
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getMap(CharTree $charTree): array
-    {
-        $codeMap = [];
-
-        // Add coded characters.
-        foreach ($this->codeMap as $length => $chars) {
-            foreach ($this->charExtractors[$length]->apply($charTree) as $char) {
-                if (isset($chars[mb_strtoupper($char)])) {
-                    $codeMap[$char] = $chars[mb_strtoupper($char)];
-                }
-            }
-        }
-
-        // Add uncoded characters.
-        foreach ($this->charExtractors[1]->apply($charTree) as $char) {
-            $codeMap[$char][] = $char;
-        }
-
-        return $codeMap;
     }
 
     /**