+++ tests/CodeMap/UpperCaseMapTest.php
@@ -11,65 +11,9 @@ use Stadly\PasswordPolice\CharTree;
  * @coversDefaultClass \Stadly\PasswordPolice\CodeMap\UpperCaseMap
  * @covers ::<private>
  * @covers ::<protected>
- * @covers ::__construct
  */
 final class UpperCaseMapTest extends TestCase
 {
-    /**
-     * @covers ::getMap
-     */
-    public function testCanGetCodeMap(): void
-    {
-        $codeMap = new UpperCaseMap();
-
-        $charTree = CharTree::fromArray([
-            CharTree::fromString('fOo'),
-            CharTree::fromString('BaR'),
-            CharTree::fromString('Baz'),
-            CharTree::fromString('123'),
-            CharTree::fromString('áÁ'),
-            CharTree::fromString('æ'),
-            CharTree::fromString('Øô'),
-            CharTree::fromString('Ëñ'),
-        ]);
-        self::assertEquals([
-            'f' => ['F'],
-            'B' => ['B'],
-            '1' => ['1'],
-            'á' => ['Á'],
-            'æ' => ['Æ'],
-            'Ø' => ['Ø'],
-            'Ë' => ['Ë'],
-        ], $codeMap->getMap($charTree), '', 0, 10, true);
-    }
-
-    /**
-     * @covers ::getMap
-     */
-    public function testCanGetCodeMapForEmpty(): void
-    {
-        $codeMap = new UpperCaseMap();
-
-        $charTree = CharTree::fromArray([
-        ]);
-        self::assertEquals([
-        ], $codeMap->getMap($charTree), '', 0, 10, true);
-    }
-
-    /**
-     * @covers ::getMap
-     */
-    public function testCanGetCodeMapForEmptyString(): void
-    {
-        $codeMap = new UpperCaseMap();
-
-        $charTree = CharTree::fromArray([
-            CharTree::fromString(''),
-        ]);
-        self::assertEquals([
-        ], $codeMap->getMap($charTree), '', 0, 10, true);
-    }
-
     /**
      * @covers ::getLengths
      */